feat(web): move initial viewport settings to extended settings
